+++ src/LuminoEngine/src/Graphics/RHIs/DirectX12/DX12Helper.hpp
@@ -57,7 +57,7 @@ public:
 	static size_t alignUpWithMask(size_t value, size_t mask) { return ((size_t)value + mask) & ~mask; }
 	static size_t alignUp(size_t value, size_t alignment = Alignment) { return alignUpWithMask(value, alignment - 1); }
 
-	static size_t getFormatSize(DXGI_FORMAT value);
+	static int32_t getFormatSize(DXGI_FORMAT value);
 	static bool getDrawPrimitiveData(PrimitiveTopology primitive, int primitiveCount, D3D_PRIMITIVE_TOPOLOGY* outTopology, UINT* outVertexCount);
 
 	static const char* LNVertexElementUsageToSemanticName(VertexElementUsage value);
